pyproject.toml
[tool.poetry]
name = "google-calendar-notifications"
version = "0.1.0"
description = ""
authors = ["Sebastian Marines <sebastian0marines@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.19.1"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
gcsa = "^1.2.0"
notify2 = "^0.3.1"
dbus-python = "^1.2.16"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.910"
yapf = "^0.31.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
